requirements.txt
asgiref
beautifulsoup4
bootstrap4
cffi
cryptography
defusedxml
diff-match-patch
Django
django-bootstrap4
django-debug-toolbar
django-import-export
django-misaka
django-session-security
django-tinymce4-lite
et-xmlfile
houdini.py
jdcal
jsmin
MarkupPy
misaka
odfpy
openpyxl
pycparser
pydot
pygal
Pygments
pyparsing
python-magic
pytz
PyYAML
six
soupsieve
sqlparse
tablib
xlrd
xlwt
